requirements.txt
Flask
python-dotenv
urllib3
TheHive4py
pymisp
